implementing new function decodeInputByTx
Function decodeInputByTx it's the better way when you need to analyse more than 200 transaction IDs per second and not wish to get ratelimited on trongrid.io.
This function doesn't need to get the contract ABI on /wallet/getcontract endpoint.
If you wish to get contract functions like Transfer, Transferfrom and balanceOf, just use the ABI I've defined on this file. That's enough for my for this moment.